// src/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json }
  | Json[]

export interface Database {
// ...
  public: {
    Tables: {
      permissions: {
        Row: {
          created_at: string | null
          id: string
          user_id: string | null
        }
        Insert: {
          created_at?: string | null
          id?: string
          user_id?: string | null
        }
        Update: {
          created_at?: string | null
          id?: string
          user_id?: string | null
        }
      }
      storefront_categories: {
        Row: {
          created_at: string | null
          icon: string | null
          id: string
          is_general: boolean
          title: string | null
        }
        Insert: {
          created_at?: string | null
          icon?: string | null
          id?: string
          is_general?: boolean
          title?: string | null
        }
        Update: {
          created_at?: string | null
          icon?: string | null
          id?: string
          is_general?: boolean
          title?: string | null
        }
      }
      storefront_images: {
        Row: {
          created_at: string | null
          credits: string | null
          description: string | null
          id: string
          image_url: string | null
          storefront_id: string | null
        }
        Insert: {
          created_at?: string | null
          credits?: string | null
          description?: string | null
          id?: string
          image_url?: string | null
          storefront_id?: string | null
        }
        Update: {
          created_at?: string | null
          credits?: string | null
          description?: string | null
          id?: string
          image_url?: string | null
          storefront_id?: string | null
        }
      }
      storefronts: {
        Row: {
          category: string | null
          created_at: string | null
          description: string | null
          id: string
          location: Json
          owner_id: string | null
          published: boolean
          title: string
        }
        Insert: {
          category?: string | null
          created_at?: string | null
          description?: string | null
          id?: string
          location: Json
          owner_id?: string | null
          published?: boolean
          title?: string
        }
        Update: {
          category?: string | null
          created_at?: string | null
          description?: string | null
          id?: string
          location?: Json
          owner_id?: string | null
          published?: boolean
          title?: string
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      is_admin: {
        Args: Record<PropertyKey, never>
        Returns: boolean
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}